Rename src folder to hp_ai to match standarp python practice
