+++ hp_ai/openai_client.py
