+++ tests/test_openai_client.py
@@ -11,7 +11,7 @@ from openai import NotFoundError
 # Add the src directory to the path so we can import the module
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-from src.openai_client import OpenAIClient
+from hp_ai.openai_client import OpenAIClient
 
 @openai_responses.mock()
 @patch.dict(os.environ, {"OPENAI_API_KEY": ""})
